Remove <leader> override and map <space> to <leader>.
Until now I overrode the leader with ',' which is common practice. It
has two drawbacks, thought. It makes ',' original purpose inaccessible
and '\' can't be used as leader anymore. Solve this by mapping <space>
to <leader>. <space> is more or less unused (mapped to move right) and
the mapping ensures that '\' is still usable as leader.