Add 'FastFold', a plugin to speed up automatic foldings.
Vim updates all kind fo automatic folds too soon and too often. This
slows the editor considerably down, especially when using the 'expr'
fold method. When the files become larger, even fast fold expressions
become a performance hook. To the point that the editor starts to lag.
This plugin switches the fold method to manual and explicitly calls
the falls expression when necessary. On a large Markdown document with
~56.000 words in ~7400 lines this makes a difference in CPU utilization
while typing of factor 10.