Add a forground color for TabLineFill to the PaperColor theme.
The TabLineFill highlight group is, as the name says, intended as a fill
color for the tabbar. But the WinTabs plugin abuses the tabbar and
writes the names of inactive tabs into the fill. Since PaperColor set
the foreground color to the same color as the background color, these
names are invisible. Fix that by changing the foreground color to
something sane.

Following the inner logic of the colorscheme, the optimal color value
would be #64646, 27 LHc lightness steps above the background. Since that
color isn't part of the 256 color terminal palette, choose #6c6c6c.