Work around bug (race?) when resetting the filetype.
Without this work around yuggdrasil resets the filetype unconditionally
whenever the tree is altered. Whith the last update of vim-lsp this
exposed a problem: There seems to be a small window where the filetype
is not set correctly, giving vim-lsp a chance to process the file with
ccls and - of couse - to highlight a lot of false errors in it.

Work around this by setting the filetype only if it's not already
yggdrasi.l. I'll need to work with a few days to see if it's save
before I will it send upstream.