Override spellchecker highlight groups to support undercurls.
Many themes have support for undercurls when running under GUI but not
when running in modern terminals. This is caused by the fact that the
theme author has no reliable way to determine if the terminal has
support for undercurls or not. Relying on fallbackl magic is error
prone... Since we explicitly configure some terminals and can be sure
(enough) that they support undercurls, override the spellchecker
highlight groups accordingly.

This enables undercurls for the spellchecker for all installed themes,
including the bundled ones.