Make the `in_visual` option in vimbits HighlightOnYank actually working.
The upstream implementation doesn't work, because `visualmode` isn't
set when the TextYankPost autocmd is called from visual mode. It's
better to rely on the `visual` property off the autocmd event. Also
do not mess with the internal `visualmode` state, that may have
unintended side effects.

I'll use this for a few days before I'll send it upstream.