Bite the bullet and finally add the vim-tmux-mnavigator.
My workflow has changed over the last year, mostly by always working in
tmux and never outside of it. It became more and more obvious that a
better integration between vim and tmux would be nice. Implement this
though the vim-tmux-navigator.

Other than the default config this used Meta instead of Ctrl a modifier
to limit the impact on existing key bindings.