Replace the venerable CtrlP fuzzy finder with the more modern Fuzzyy.
CtrlP has been unmaintained for years. While it just works, it's written
in legacy Vimscript and reimplements functionality in it which has been
provided by the core editor for a long time. Fuzzyy is an alternative
written in modern Vim9. It's much faster, especially on big directories,
has a modern look and feel and some nice to have features.