Enable list mode by default and configure it to show tabs and spaces.
This is a kind of experiment. I might keep it or not. Thing is, that I
had `visualedit=all` since forever and disabled it late last year due to
it's bugs with virtual text and conceal. This (re)exposed Vims annoying
behavior of position the cursor on the last character of a tab. That can
be hacked around by setting `listchars=\ \ `. But when I set listchars
and enable list mode anyways, I can use it for something more useful.