Don't wrap cursor shape changes into Ptmux escape sequences.
tmux has already for some time tracked cursor shape changes by itself
and no longer delegates them to the underlying terminal. When the
corresponding escape sequences are wrapped into Ptmux sequences, tmux
ignores them. That breaks tmux internal state tracking.  Cursor changes
are either ignores or we are ending up with the wrong shape.

This is arguably a hack. Terminus is one of these plugins which have
been more or less overcome by Vim gaining the same functionality in the
core. The same goes for autocompletion and maybe fuzzy finding. The
bigger plan is to let things settle for a few month and afterward have
a look which plugins can be removed.