Move fuzzy to plugins/.
I've pondered a little bit if I shpuld keep fuzzy or move to simpler
fuzzy finding logic directly in vimrc. I decided to keep fuzzy, because
it's a small plugin (just above 3000 lines of code) with great
functionality and speed. It's hard to come up with something that
is "better" in whatever regard.