Remove the supertab and vimcomplete plugins.
These will be replaced Vims brand new integrated completion functions.
These allow to define custom completors right in the core editor without
the need of complex plugins.

Supertab was one of my first plugins. It was overcome by the advent of
modern completion plugins and only kept because it just worked(tm). It's
time to let it go...