Implement a completion function for the lsp plugin.
This completion function takes the omnifunc of the lsp plugin as data
source and feeds the generated suggestion info Vims internal completion
engine. This isn't that different from the vimcomplete plugin. In fact
the code is heavily based on vimcompletes lsp completor. But it uses
the completion engine of the core editor, which is faster.