Replace fuzzyy by ctrlp, revert 6141557.
About two weeks ago I wrote, that I pondered for some time which fuzzy
finder I want to use and that I decided to go with fuzzyy. fuzzyy is
nice. It's small, it's modern with it's use of popups and the preview
functionality and it's written in vim9script. However it has some bugs
which only became apparent while testing the changes of the last week.
The worst one is, that opening a file with fuzzyy causes state to leak
from the new buffer into the one where fuzzy was triggered from. For
example the LSP state of the buffers isn't isolated, causing strange
problems. I skipped through the code, however the problem isn't obvious
and I don't have the motivation to fix it. And such problem shows, that
fuzzyys userbase is too small. Such problem shouldn't go unnoticed for
month, if not years... Therefor bite the bullet and revert back to ctrlp
for the time being. It's not modern, but reliable. It has worked for
many years and will do that for the years to come.