Switch to build-in autocompletion.
Vim gained build-in autocompletion in 9.1.1590. From a functional point
of view it's the same as we did in vim9script so far, just (at least in
theory) faster because it's part of the core. It also covers some corner
cases. For example very long response times by some completion sources.
And it further simplifies the configuration.

Bumps the minimum required Vim version to 9.1.1590.